living-twin-synthetic-data/pyproject.toml
[project]
name = "living-twin-synthetic-data"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "beautifulsoup4>=4.13.5",
    "boto3>=1.40.29",
    "click>=8.2.1",
    "faker>=37.6.0",
    "jsonschema>=4.25.1",
    "openai>=1.107.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "rich>=14.1.0",
]
